[Refactor/#496] QuestListViewController 에서 컬렉션뷰의 역할 분리
